Adds Brandon thermostate
In this case, we just do the lennard-Jones potential. But we are keeping temperature (Kinetic Energy) fixed. In every step when the Kinetic Energy changes, we scale the velocity and force the maen Kinetic Energy per particle to be fixed! hence we see the by controlling the temperature we can actually observe the phase changes in them! The Crystall pattern can easily observed!